app/backend/requirements.txt
fastapi == 0.121.0
uvicorn[standard] == 0.38.0
sqlalchemy == 2.0.44
argon2-cffi == 25.1.0
python-jose[cryptography] == 3.5.0
python-multipart == 0.0.20
pydantic == 2.12.4
pydantic-settings == 2.11.0
python-dotenv == 1.2.1
google-generativeai == 0.8.5
email-validator == 2.3.0
websockets == 15.0.1
openpyxl == 3.1.5
reportlab == 4.4.4
psycopg2-binary == 2.9.10

pytest == 9.0.1
httpx == 0.28.1
pytest-cov == 7.0.0
